fix: Address publisher metrics compiler warnings from TiCS
