+++ static/js/publisher/pages/Metrics/metrics/config.ts
