+++ static/js/publisher/pages/Metrics/metrics/graphs/activeDevicesGraph/index.ts
@@ -8,7 +8,6 @@ import {
   prepareStackedData,
   prepareLineData,
   prepareScales,
-  prepareAnnotationsData,
   prepareAxis,
 } from "./dataProcessing";
 import {
