--- static/js/publisher/pages/Metrics/metrics/graphs/territories.js
@@ -1,5 +1,0 @@
-import map from "../../../../../public/snap-details/map";
-
-export default function (el, data) {
-  map(el, data);
-}
