+++ static/js/publisher/pages/Metrics/metrics/graphs/territories.ts
@@ -1,0 +1,7 @@
+import map from "../../../../../public/snap-details/map";
+
+import type { TerritoriesMetricsData } from "../../../../types/shared";
+
+export default function (el: string, data: TerritoriesMetricsData) {
+  map(el, data);
+}
